+++ CloudShell.ResourceDefinitions.ReferenceProviders/SecretsVault/SecretsVaultResourceTypeServiceCollectionExtensions.cs
@@ -45,7 +45,7 @@ public static class SecretsVaultResourceTypeServiceCollectionExtensions
         services.TryAddSingleton<SecretsVaultProcessRuntimeController>();
         services.TryAddSingleton<ISecretsVaultRuntimeController>(
             serviceProvider => serviceProvider.GetRequiredService<SecretsVaultProcessRuntimeController>());
-        services.TryAddSingleton<ISecretsVaultInspector, NoopSecretsVaultInspector>();
+        services.TryAddSingleton<ISecretsVaultInspector, SecretsVaultRuntimeInspector>();
 
         return services;
     }